Remove passthrough asserts in prepare_output_buffer
base_transform_prepare_output_buffer asserts we are not in
passthrough mode when we are returning a buffer which is not
the original inputbuffer.

This assertion is not correct. In the original default_generate_output
the passthrough mode is only checked after prepare_output_buffer
not before unlike the Rust bindings.
As stated in the documentation
https://gstreamer.freedesktop.org/documentation/base/gstbasetransform.html?gi-language=c#modifications-only-to-the-capsmetadata-of-a-buffer
when only modifying the metadata of a buffer we do not require a
writeable buffer and should subbuffer the original. This should be
possible in passthrough mode so we don't need to call a
transform_function as we do not edit the buffer data but do need
to copy it to create a subbuffer. The assertions prevent this so they
have been removed.